Core (9 files) — int → decimal on every quantity field/parameter:   - PositionData.CurrentQuantity, PositionInfo.Quantity, OrderInfo.Quantity/FilledQuantity,   OrderIntentEvent.Quantity, OrderUpdateEvent.FilledQuantity/RemainingQuantity   - IBrokerService, IOrderManager, IPositionTracker, IStateRepository (including   OrderIntentDto.Quantity)
  Infrastructure (4 files):
  - Entities.cs — 6 quantity properties changed to decimal
  - EntityConfigurations.cs — added HasPrecision(18, 8) for each quantity column
  - AlpacaBrokerService — p.Quantity (positions) no longer cast; orders use
  (decimal?)order.IntegerQuantity ?? 0m with a comment noting the SDK v7.2.0 limitation
  (IntegerQuantity is null for fractional orders — the broker already submits via
  OrderQuantity.Fractional() so orders reach Alpaca correctly, but the read-back quantity from
  Alpaca responses won't reflect the fractional value until the SDK is upgraded or raw JSON is
  parsed)
  - StateRepository — 3 signature changes + return tuple type

  Trading (3 files):
  - PositionSizer — both overloads gain bool allowFractional = false; crypto path skips floor and
  uses Math.Max(0.0001m, Math.Round(maxQty, 8))
  - OrderManager — detects crypto from options.Symbols.CryptoSymbols, passes allowFractional to
  sizer, handles drawdown multiplier rounding per asset class
  - PositionTracker — signature update

  Worker (3 files):
  - EventDispatcherService — exit handler now reads actual position qty from tracker instead of
  the 100 placeholder; signal handler passes 0m sentinel
  - StreamPollerService — Math.Max(0m, ...) for decimal remainder
  - ReconciliationService — tuple type updated

  Migration — Migrations/FractionalQuantity/ alters 6 columns (order_intents, fills,
  position_tracking, position_snapshots, trades) from INTEGER to TEXT with precision 18,8. SQLite
  reads existing integer rows transparently.

  Tests — all Arg.Any<int>() and Arg.Is<int>() matchers updated to decimal equivalents.

+++ cs/src/AlpacaFleece.Core/Interfaces/IBrokerService.cs
@@ -28,7 +28,7 @@ public interface IBrokerService
     ValueTask<OrderInfo> SubmitOrderAsync(
         string symbol,
         string side,
-        int quantity,
+        decimal quantity,
         decimal limitPrice,
         string clientOrderId,
         CancellationToken ct = default);
